work, more security, added user to reviews, etc.

+++ api/src/main/java/com/example/review/ReviewService.java
@@ -4,9 +4,12 @@ import com.example.movie.Movie;
 import com.example.movie.MovieRepository;
 import com.example.movie.exception.MovieNotFoundException;
 import com.example.review.exception.ReviewNotFoundException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import javax.ws.rs.NotAuthorizedException;
 
 @Service
 public class ReviewService {
@@ -19,30 +22,50 @@ public class ReviewService {
         this.movieRepository = movieRepository;
     }
 
-    public List<Review> getReviewsByMovieId(Integer id) {
-        if (!movieRepository.existsById(id)) {
-            throw new MovieNotFoundException(id);
-        }
-
-        return reviewRepository.findByMovieId(id);
-    }
-
     public Review getReview(Integer id) {
         return reviewRepository.findById(id).orElseThrow(() -> new ReviewNotFoundException(id));
     }
 
-    public Review createReview(Review review, Integer id) {
-        Movie movie = movieRepository.findById(id)
-                .orElseThrow(() -> new MovieNotFoundException(id));
-        movie.getReviews().add(review);
+    public Review createReview(Review review, Integer movieId) {
+        Movie movie = movieRepository.findById(movieId)
+                .orElseThrow(() -> new MovieNotFoundException(movieId));
+        String userName = getUserNameFromContext();
         review.setMovie(movie);
+        review.setUserName(userName);
         return reviewRepository.save(review);
     }
 
+    public Review updateReview(Review review) {
+        String userName = getUserNameFromContext();
+
+        if (userName.equals(review.getUserName())) {
+            return reviewRepository.save(review);
+        }
+
+        throw new NotAuthorizedException("You can update only your reviews");
+    }
+
     public void deleteReview(Review review) {
-        Movie movie = movieRepository.findById(review.getId())
-                .orElseThrow(() -> new MovieNotFoundException(review.getId()));
-        movie.getReviews().remove(review);
-        reviewRepository.delete(review);
+        String userName = getUserNameFromContext();
+        boolean isAdmin = getIsAdminFromContext();
+
+        if (userName.equals(review.getUserName()) || isAdmin) {
+            reviewRepository.delete(review);
+            return;
+        }
+
+        throw new NotAuthorizedException("You can delete only your reviews");
+    }
+
+    private String getUserNameFromContext() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getName();
+    }
+
+    private boolean getIsAdminFromContext() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .anyMatch(role -> role.equals("ROLE_admin"));
     }
 }
